5> Remove #FlowFixMe for get edge method
Summary:
For get edge function,
we use to have function like this
```
 getCampaigns (fields: Array<string>, params: Object = {}, fetchFirstPage: boolean = true): Campaign {
    // $FlowFixMe : Support Generic Types
    return this.getEdge(
      ...
    );
  }
```
Which is not right.

And in `AbstractCrudObject`
getEdge function should be able to return `Cursor ` or ` Promise<*>` depends on the param `fetchFirstPage`.

So in this diff, firstly, I change `AbstractCrudObject` -> return  `Cursor | Promise<*>`
Then for all objects class get edge, return  ``Cursor | Promise<*>``.
Then we can remove the #FlowFixMe for get edge calls.

Reviewed By: codytwinton

Differential Revision: D15708622

fbshipit-source-id: 5ffaf735